pivot_confluence_strategy_corrected/requirements.txt
pandas==2.3.2
numpy==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
alpaca-trade-api==3.2.0
tabulate==0.9.0
python-dotenv==1.2.1
requests==2.32.5
tzdata>=2022.7
